vesuvius-competition/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
opencv-python>=4.8.0
albumentations>=1.3.1

# Model architectures
segmentation-models-pytorch>=0.3.3
timm>=0.9.0

# Data handling
pandas>=2.0.0
h5py>=3.8.0
zarr>=2.15.0
tifffile>=2023.7.0
Pillow>=10.0.0

# Training utilities
wandb>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
rich>=13.0.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
jupyter>=1.0.0
ipython>=8.14.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0

# Competition specific
kaggle>=1.5.0
modal>=0.50.0
